/login bugs- any login combination will get you to the home page: fixed with sending the correct response from the server and then skipping the cookie creation on unsucessfull loging
/habits/get bugs - cards not rendering and some values were blank:
incorrect data structure was being sent back. changed to an array of objects
having it fetch data constantly so it get updated. *stretch it to save the fetch data in a cache and only update when habits are added.

close form upon submission

serverside to provide the html file on any url address. the react-router will handle incorrect paths.